requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.13.0
gspread>=5.7.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
altair>=5.0.0
garminconnect>=0.1.50
